json: add Value[T] and ByteString
Value[T] allows full-fidelity decoding of struct fields that may be
undefined or null.

If a struct field of type Value[T] has the zero value and is marked
omitempty, it is omitted. Otherwise, it is marshaled as the null value.
A Value[T] that is not a struct field and is undefined or null is
marshaled as the null value. A Value[T] that is defined and non-null
is marshaled as the representation of its Value.

During unmarshaling, struct fields of type Value[T] that have no
corresponding property P in the input are assigned the zero value (and are
therefore undefined). If P is the null value, the value of the field is
null. If P is non-null, the Value[T] field will attempt to unmarshal P
as a value of type T.

ByteString allows byte slices to be treated as strings for the purposes
of marshaling and unmarshaling. This can be useful when marshaling text
that is otherwise represented as bytes.